Adding new licensed images and refactoring lab3 advanced game table layout
