+++ client/src/assets/images/landingpage/lab_thumbnails/ear.jpg
