+++ client/src/assets/images/landingpage/landingpage_hero.jpg
