+++ client/src/assets/images/landingpage/map-image.png
