+++ client/src/assets/stylesheets/components/App.scss
@@ -137,7 +137,7 @@ header.masthead {
 	text-align: center;
 	color: white;
 	//height: auto;
-	background-image: url("../../images/div_head.jpg") !important;
+	background-image: url("../../images/landingpage/landingpage_hero.jpg") !important;
 	background-repeat: no-repeat;
 	background-attachment: scroll;
 	background-position: center center;
@@ -690,7 +690,7 @@ a.icon {
 }
 section#contact {
 	background-color: #212529;
-	background-image: url(../../images/map-image.png);
+	background-image: url(../../images/landingpage/map-image.png);
 	background-repeat: no-repeat;
 	background-position: center;
 	.section-heading {
