+++ client/src/components/exercise/lab3/pages/BeginnerExercise/Exercise.js
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
-import catImage from "../../../../../assets/images/c1.svg";
-import carImage from "../../../../../assets/images/c2.svg";
-import burgerImage from "../../../../../assets/images/b.png";
-import cowImage from "../../../../../assets/images/cow.jpg";
+import catImage from "../../../../../assets/images/lab3/exercise/cat.svg";
+import carImage from "../../../../../assets/images/lab3/exercise/car.svg";
+import burgerImage from "../../../../../assets/images/lab3/exercise/hamburger.svg";
+import cowImage from "../../../../../assets/images/lab3/exercise/cow.svg";
 import CatClickFirstNavigate from "../../helpers/CatClickFirstNavigate";
 import { Typography } from "@material-ui/core";
 import {PageService} from "../../../../../services/PageService";
@@ -95,7 +95,7 @@ class Exercise extends Component {
             <tr>
               <td>
                 <button style={imgStyle} onClick={() => catClick()}>
-                  <img src={catImage} alt={"image1"} />
+                  <img style={{ transform: "scale(1.25)"}} src={catImage} alt={"image1"} />
                 </button>
               </td>
               <td>
