+++ client/src/components/header/header.js
@@ -1,5 +1,5 @@
 import React, {useEffect, useState} from 'react';
-import Logo from "../../assets/images/ALL_Logo_Header.svg";
+import Logo from "../../assets/images/logos/ALL_Logo_Header.svg";
 import "../../assets/stylesheets/components/Header.scss"
 import WelcomeMessage from './helpers/WelcomeMessage';
 import {connect} from "react-redux";
